Tests Fixed:   - 81757: Port isolation with UDN and bridge CNI - NOW PASSING   - 80526: UDN with mixed port isolation - NOW PASSING   - 69947: Macvlan Unsolicited NA - AWS skip logic (already working)
  Tests Documented (require infrastructure changes):
  - 67625: ovnkube-trace pod-to-pod - RBAC permissions needed
  - 67648: ovnkube-trace pod-to-hostnet - RBAC permissions needed

  Key Changes for UDN Tests:

  1. IPv4-only subnet configuration (cluster doesn't support IPv6):
     - Changed from dual-stack to single subnet: 10.132.0.0/16
     - Matches scale team's working configuration
     - Fixes: cluster does not support ipv6 controller errors

  2. Correct UDN status condition check:
     - Changed from NetworkReady to NetworkCreated
     - This is the actual condition type used by UDN controller
     - Fixes: 180-second timeout waiting for wrong condition

  3. Added required namespace labels:
     - k8s.ovn.org/primary-user-defined-network: layer2-ipv4-pudn1
     - security.openshift.io/scc.podSecurityLabelSync: false
     - Per Arti's guidance and scale team patterns

  4. Increased timeout from 120s to 180s for slower clusters

  RBAC Documentation (67625, 67648):

  Added detailed FIXME comments explaining:
  - ovnkube-trace needs pods/exec permission
  - Current ClusterRole has pods/nodes get/list but missing pods/exec
  - Cannot add temporary RBAC in tests (security risk in CI)
  - Tests marked [informing] so won't block CI
  - Requires team decision on proper RBAC solution
