+++ openshift/test/e2e_test.go
@@ -4,6 +4,9 @@ import (
 	// import OVN-Kubernetes E2Es
 	_ "github.com/ovn-kubernetes/ovn-kubernetes/test/e2e"
 
+	// import OVN-Kubernetes OTP tests
+	_ "github.com/ovn-kubernetes/ovn-kubernetes/openshift/test/otp"
+
 	// Ensure that logging flags are part of the command line.
 	_ "k8s.io/component-base/logs/testinit"
 )
